ITunesIOSAssetExporter: Export synchronously for now
This is a bit of a workaround, since Xcode will rightfully warn about
having an interactive thread wait on a low-priority thread here. Fixing
this would require rearchitecting the location checks in
BaseExternalTrackModel etc, so this is more convenient for now, even
though it comes at a risk of minor UI hangs when selecting tracks from
an external source such as iTunes.